Merge commit '22526e698d410058e24987164a4e6a01bd4d80b0'

* commit '22526e698d410058e24987164a4e6a01bd4d80b0':
  scroll text field with touch

+++ WebKit/android/jni/WebViewCore.h
@@ -319,7 +319,7 @@ namespace android {
         /**
          * Scroll the focused textfield to (x, y) in document space
          */
-        void scrollFocusedTextInput(int x, int y);
+        void scrollFocusedTextInput(float x, int y);
         void setFocusControllerActive(bool active);
 
         void saveDocumentState(WebCore::Frame* frame);
